requirements.txt
beautifulsoup4==4.13.3
dvc-s3==3.2.0
fastapi==0.115.12
joblib==1.4.2
mlflow==2.11.3
numpy==1.24.4
pandas==2.2.3
pycryptodome==3.22.0
psycopg2-binary==2.9.9
pytest==8.3.5
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.1
sqlalchemy==2.0.29
uvicorn==0.34.0y
